Map Display Plugin for MyCampusEAI
ViewController.h and .m controls the basic functionality of displaying
map, plotting points and polylines on Map and communicating to the
google Map API.
MyLocation.h and .m is a custom made annotation.
Waypoint.h and .m is an object to hold the lat, long and details of a
particular point.
MapView.xib is the view containing the mapView